app/__init__.py
from flask import Flask
from dotenv import load_dotenv
import os

# Load environment variables
load_dotenv()

def create_app():
    app = Flask(__name__)
    app.config['SECRET_KEY'] = os.getenv('SECRET_KEY')

    # Register blueprints
    from .routes import main_routes
    app.register_blueprint(main_routes)

    return app