requirements.txt
Flask==2.3.2
gunicorn==20.1.0
pytube==15.0.0
moviepy==1.0.3
Pillow==10.0.0
instaloader==4.9.2
TikTokApi==5.1.0
spotdl==4.1.0
python-dotenv==1.0.0
opencv-python==4.7.0.72
torch==2.0.1
torchvision==0.15.2
numpy==1.23.5
Pillow==10.0.0